begin extracting database-related functionality
